// src/app/feature/entities/matching-version/models/matching-editor.models.ts
export interface VersionCardData {
  id: number;
  version: number;
  createdAt: Date;
  status: string;
  matchingStats: MatchingStats;
  canEdit: boolean;
  canDelete: boolean;
  canExport: boolean;
}

export interface MatchingStats {
  totalEvents: number;
  matchedCount: number;
  unmatchedCount: number;
  conflictCount: number;
  matchedPercentage: number;
  unmatchedPercentage: number;
  conflictPercentage: number;
}